requirements.txt
requests
pandas
python-dotenv
beautifulsoup4